# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Playero API",
        "description": "API oficial do sistema Playero",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/v1/agenda/config": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Obter configuracao da agenda",
                "operationId": "ac33d380c93f8d1dda870cb0b6079f0c",
                "responses": {
                    "200": {
                        "description": "Configuracao atual",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Configuracao da agenda carregada com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "hora_abertura": {
                                                    "type": "string",
                                                    "example": "08:00"
                                                },
                                                "hora_fechamento": {
                                                    "type": "string",
                                                    "example": "22:00"
                                                },
                                                "duracao_reserva_minutos": {
                                                    "type": "integer",
                                                    "example": 60
                                                },
                                                "dias_semana_ativos": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "America/Sao_Paulo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Configuracao nao encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Criar ou atualizar configuracao da agenda",
                "operationId": "2e6537ab7174903aebaaf1f0be61bbe2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hora_abertura",
                                    "hora_fechamento",
                                    "duracao_reserva_minutos",
                                    "dias_semana_ativos"
                                ],
                                "properties": {
                                    "hora_abertura": {
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "hora_fechamento": {
                                        "type": "string",
                                        "example": "22:00"
                                    },
                                    "duracao_reserva_minutos": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "dias_semana_ativos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "America/Sao_Paulo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configuracao atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Configuracao da agenda atualizada com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "hora_abertura": {
                                                    "type": "string",
                                                    "example": "08:00"
                                                },
                                                "hora_fechamento": {
                                                    "type": "string",
                                                    "example": "22:00"
                                                },
                                                "duracao_reserva_minutos": {
                                                    "type": "integer",
                                                    "example": 60
                                                },
                                                "dias_semana_ativos": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "America/Sao_Paulo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Configuracao criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Configuracao da agenda criada com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "hora_abertura": {
                                                    "type": "string",
                                                    "example": "08:00"
                                                },
                                                "hora_fechamento": {
                                                    "type": "string",
                                                    "example": "22:00"
                                                },
                                                "duracao_reserva_minutos": {
                                                    "type": "integer",
                                                    "example": 60
                                                },
                                                "dias_semana_ativos": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "America/Sao_Paulo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/agenda/blockings/{id}": {
            "delete": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Excluir bloqueio",
                "operationId": "99493aab3e2a7ceb9eb3c1c90542e861",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bloqueio removido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bloqueio removido com sucesso."
                                        },
                                        "data": {
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bloqueio nao encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}